+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.1.1
+- Disabled the second `*/` on class definitions if that setting is false
+- Adds class definitions to the right-click context action, however they are not whitespaced with the class definition like if using the snippet feature
+
 ## 1.1.0
 - Added new option to disable the second `*/` that shows up in some VSCode installations